--- KeepFresh/SearchTab.swift
@@ -1,8 +1,0 @@
-import SwiftUI
-
-struct SearchTab: View {
-    var body: some View {
-        Text("SearchTab")
-//        SearchView()
-    }
-}
