+++ Packages/Features/Package.swift
@@ -10,13 +10,13 @@ let package = Package(
         .library(
             name: "SearchUI",
             targets: ["SearchUI"]
-        ),
+        )
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
         // Targets can depend on other targets in this package and products from dependencies.
         .target(
             name: "SearchUI", dependencies: []
-        ),
+        )
     ]
 )
